+++ changelog.md
@@ -2,6 +2,7 @@
 - Allowed string lists to be modifiable via the config screen
 - Added a config option to stop getting OpenGL messages
 - Fixed config comments not being displayed
+- Added a reset button to each config entry
 
 [5.0.0]
 - Made initial port to 1.18